data_load/requirements.txt
numpy==2.4.4
pandas==3.0.2
python-dotenv==1.2.2
